Finally got rid of the uneven jittering!!
The issue was that the direction towards which I needed to round the
camera coordinates (either infinity or minus-infinity) depends on the
direction that you're moving. Once I accounted for that in a previous
commit, and once I rounded the coordinates whenever a change in
controller status occurred, the jittering went away!