I kept working on the loading system for the character's animations.
The `load_animation()` function loads successfully; I checked the
surface animations, and they did load indeed. However, the problem lies
in the returning of the function. Something is going on that the
function doesn't want to return the Animation struct. I think a possible
solution for this might be to change all references to animations to
work based on pointers.